@@ -1,4 +1,4 @@
-﻿﻿using System;
+﻿using System;
 using System.Collections.ObjectModel;
 using System.Linq;
 using System.Threading.Tasks;
@@ -14,7 +14,7 @@ namespace Arcus.Testing
     /// </summary>
     public class OnSetupTemporaryTopicSubscriptionOptions
     {
-        private readonly Collection<Action<CreateSubscriptionOptions>> _configureSubscriptionOptions = new();
+        private readonly Collection<Action<CreateSubscriptionOptions>> _configureSubscriptionOptions = [];
 
         /// <summary>
         /// Configures the <see cref="Azure.Messaging.ServiceBus.Administration.CreateSubscriptionOptions"/> used when the test fixture creates the topic subscription.
@@ -53,7 +53,7 @@ namespace Arcus.Testing
         /// <summary>
         /// Gets the options related to setting up the <see cref="TemporaryTopicSubscription"/>.
         /// </summary>
-        public OnSetupTemporaryTopicSubscriptionOptions OnSetup { get; } = new OnSetupTemporaryTopicSubscriptionOptions();
+        public OnSetupTemporaryTopicSubscriptionOptions OnSetup { get; } = new();
     }
 
     /// <summary>
@@ -62,10 +62,10 @@ namespace Arcus.Testing
     public class TemporaryTopicSubscription : IAsyncDisposable
     {
         private readonly ServiceBusAdministrationClient _client;
-        private readonly Collection<CreateRuleOptions> _rules = new();
+        private readonly Collection<CreateRuleOptions> _rules = [];
 
-        private readonly CreateSubscriptionOptions _options;
         private readonly bool _createdByUs;
+        private readonly DisposableCollection _disposables;
         private readonly ILogger _logger;
 
         private TemporaryTopicSubscription(
@@ -80,12 +80,12 @@ namespace Arcus.Testing
             ArgumentNullException.ThrowIfNull(serviceBusNamespace);
 
             _client = client;
-            _options = options;
             _createdByUs = createdByUs;
-            _logger = logger;
+            _logger = logger ?? NullLogger.Instance;
+            _disposables = new DisposableCollection(_logger);
 
-            Name = _options.SubscriptionName;
-            TopicName = _options.TopicName;
+            Name = options.SubscriptionName;
+            TopicName = options.TopicName;
             FullyQualifiedNamespace = serviceBusNamespace;
         }
 
@@ -117,9 +117,9 @@ namespace Arcus.Testing
         /// <exception cref="InvalidOperationException">
         ///     Thrown when the no Azure Service Bus topic exists with the provided <paramref name="topicName"/> in the given <paramref name="fullyQualifiedNamespace"/>.
         /// </exception>
-        public static async Task<TemporaryTopicSubscription> CreateIfNotExistsAsync(string fullyQualifiedNamespace, string topicName, string subscriptionName, ILogger logger)
+        public static Task<TemporaryTopicSubscription> CreateIfNotExistsAsync(string fullyQualifiedNamespace, string topicName, string subscriptionName, ILogger logger)
         {
-            return await CreateIfNotExistsAsync(fullyQualifiedNamespace, topicName, subscriptionName, logger, configureOptions: null);
+            return CreateIfNotExistsAsync(fullyQualifiedNamespace, topicName, subscriptionName, logger, configureOptions: null);
         }
 
         /// <summary>
@@ -138,7 +138,7 @@ namespace Arcus.Testing
         /// <exception cref="InvalidOperationException">
         ///     Thrown when the no Azure Service Bus topic exists with the provided <paramref name="topicName"/> in the given <paramref name="fullyQualifiedNamespace"/>.
         /// </exception>
-        public static async Task<TemporaryTopicSubscription> CreateIfNotExistsAsync(
+        public static Task<TemporaryTopicSubscription> CreateIfNotExistsAsync(
             string fullyQualifiedNamespace,
             string topicName,
             string subscriptionName,
@@ -148,7 +148,7 @@ namespace Arcus.Testing
             ArgumentException.ThrowIfNullOrWhiteSpace(fullyQualifiedNamespace);
 
             var client = new ServiceBusAdministrationClient(fullyQualifiedNamespace, new DefaultAzureCredential());
-            return await CreateIfNotExistsAsync(client, topicName, subscriptionName, logger, configureOptions);
+            return CreateIfNotExistsAsync(client, topicName, subscriptionName, logger, configureOptions);
         }
 
         /// <summary>
@@ -164,13 +164,13 @@ namespace Arcus.Testing
         ///     Thrown when the no Azure Service Bus topic exists with the provided <paramref name="topicName"/>
         ///     in the given namespace where the given <paramref name="adminClient"/> points to.
         /// </exception>
-        public static async Task<TemporaryTopicSubscription> CreateIfNotExistsAsync(
+        public static Task<TemporaryTopicSubscription> CreateIfNotExistsAsync(
             ServiceBusAdministrationClient adminClient,
             string topicName,
             string subscriptionName,
             ILogger logger)
         {
-            return await CreateIfNotExistsAsync(adminClient, topicName, subscriptionName, logger, configureOptions: null);
+            return CreateIfNotExistsAsync(adminClient, topicName, subscriptionName, logger, configureOptions: null);
         }
 
         /// <summary>
@@ -203,13 +203,13 @@ namespace Arcus.Testing
 
             var options = new TemporaryTopicSubscriptionOptions();
             configureOptions?.Invoke(options);
-            
+
             CreateSubscriptionOptions createOptions = options.OnSetup.CreateSubscriptionOptions(topicName, subscriptionName);
 
-            NamespaceProperties properties = await adminClient.GetNamespacePropertiesAsync();
+            NamespaceProperties properties = await adminClient.GetNamespacePropertiesAsync().ConfigureAwait(false);
             string serviceBusNamespace = properties.Name;
 
-            if (!await adminClient.TopicExistsAsync(createOptions.TopicName))
+            if (!await adminClient.TopicExistsAsync(createOptions.TopicName).ConfigureAwait(false))
             {
                 throw new InvalidOperationException(
                     $"[Test:Setup] cannot create temporary subscription '{createOptions.SubscriptionName}' on Azure Service Bus topic '{serviceBusNamespace}/{createOptions.TopicName}' " +
@@ -217,14 +217,14 @@ namespace Arcus.Testing
                     $"Please make sure to have an available Azure Service Bus topic before using the temporary topic subscription test fixture");
             }
 
-            if (await adminClient.SubscriptionExistsAsync(createOptions.TopicName, createOptions.SubscriptionName))
+            if (await adminClient.SubscriptionExistsAsync(createOptions.TopicName, createOptions.SubscriptionName).ConfigureAwait(false))
             {
-                logger.LogTrace("[Test:Setup] Use already existing Azure Service Bus topic subscription '{SubscriptionName}' in '{Namespace}/{TopicName}'", createOptions.SubscriptionName, serviceBusNamespace, createOptions.TopicName);
+                logger.LogSetupUseExistingSubscription(createOptions.SubscriptionName, serviceBusNamespace, createOptions.TopicName);
                 return new TemporaryTopicSubscription(adminClient, serviceBusNamespace, createOptions, createdByUs: false, logger);
             }
 
-            logger.LogTrace("[Test:Setup] Create new Azure Service Bus topic subscription '{SubscriptionName}' in '{Namespace}/{TopicName}'", createOptions.SubscriptionName, serviceBusNamespace, createOptions.TopicName);
-            await adminClient.CreateSubscriptionAsync(createOptions);
+            logger.LogSetupCreateNewSubscription(createOptions.SubscriptionName, serviceBusNamespace, createOptions.TopicName);
+            await adminClient.CreateSubscriptionAsync(createOptions).ConfigureAwait(false);
 
             return new TemporaryTopicSubscription(adminClient, serviceBusNamespace, createOptions, createdByUs: true, logger);
         }
@@ -234,10 +234,12 @@ namespace Arcus.Testing
         /// </summary>
         /// <param name="ruleName">The name to describe the subscription rule.</param>
         /// <param name="ruleFilter">The filter expression used to match messages.</param>
+        /// <exception cref="ObjectDisposedException">Thrown when the test fixture was already teared down.</exception>
         /// <exception cref="ArgumentNullException">Thrown when the <paramref name="ruleName"/> or <paramref name="ruleFilter"/> is <c>null</c>.</exception>
         /// <exception cref="ArgumentException">Thrown when the <paramref name="ruleName"/> represents the default rule name.</exception>
         public async Task AddRuleIfNotExistsAsync(string ruleName, RuleFilter ruleFilter)
         {
+            ObjectDisposedException.ThrowIf(_disposables.IsDisposed, this);
             ArgumentNullException.ThrowIfNull(ruleName);
             ArgumentNullException.ThrowIfNull(ruleFilter);
 
@@ -247,16 +249,16 @@ namespace Arcus.Testing
                     "Only custom Azure Service Bus topic subscription rules can be added to the test fixture, please provide a custom name for your test-managed rule", nameof(ruleName));
             }
 
-            if (await _client.RuleExistsAsync(TopicName, Name, ruleName))
+            if (await _client.RuleExistsAsync(TopicName, Name, ruleName).ConfigureAwait(false))
             {
-                _logger.LogDebug("[Test] Skip creation of Azure Service Bus topic subscription rule '{RuleName}' in '{Namespace}/{TopicName}/{SubscriptionName}' as it already exists", ruleName, FullyQualifiedNamespace, TopicName, Name);
+                _logger.LogSkipCreateSubscriptionRule(ruleName, FullyQualifiedNamespace, TopicName, Name);
             }
             else
             {
-                _logger.LogDebug("[Test] Create new Azure Service Bus topic subscription rule '{RuleName}' in '{Namespace}/{TopicName}/{SubscriptionName}'", ruleName, FullyQualifiedNamespace, TopicName, Name);
+                _logger.LogCreateSubscriptionRule(ruleName, FullyQualifiedNamespace, TopicName, Name);
 
                 var options = new CreateRuleOptions(ruleName, ruleFilter);
-                await _client.CreateRuleAsync(TopicName, Name, options);
+                await _client.CreateRuleAsync(TopicName, Name, options).ConfigureAwait(false);
 
                 _rules.Add(options);
             }
@@ -268,32 +270,73 @@ namespace Arcus.Testing
         /// <returns>A task that represents the asynchronous dispose operation.</returns>
         public async ValueTask DisposeAsync()
         {
-            await using var disposables = new DisposableCollection(_logger);
-
-            if (await _client.SubscriptionExistsAsync(_options.TopicName, _options.SubscriptionName))
+            if (_disposables.IsDisposed)
             {
-                if (_createdByUs)
-                {
-                    disposables.Add(AsyncDisposable.Create(async () =>
-                    {
-                        _logger.LogDebug("[Test:Teardown] Delete Azure Service Bus topic subscription '{SubscriptionName}' in '{Namespace}/{TopicName}'", _options.SubscriptionName, FullyQualifiedNamespace, _options.TopicName);
-                        await _client.DeleteSubscriptionAsync(_options.TopicName, _options.SubscriptionName);
-                    }));
-                }
-                else
-                {
-                    disposables.AddRange(_rules.Select(r => AsyncDisposable.Create(async () => 
-                    {
-                        if (await _client.RuleExistsAsync(TopicName, Name, r.Name))
-                        {
-                            _logger.LogDebug("[Test:Teardown] Delete Azure Service Bus topic subscription rule '{RuleName}' in {Namespace}/{TopicName}/{SubscriptionName}", r.Name, FullyQualifiedNamespace, _options.TopicName, _options.SubscriptionName);
-                            await _client.DeleteRuleAsync(TopicName, Name, r.Name);
-                        }
-                    })));
-                }
+                return;
             }
 
-            GC.SuppressFinalize(this);
+            await using (_disposables.ConfigureAwait(false))
+            {
+                if (await _client.SubscriptionExistsAsync(TopicName, Name).ConfigureAwait(false))
+                {
+                    if (_createdByUs)
+                    {
+                        _disposables.Add(AsyncDisposable.Create(async () =>
+                        {
+                            _logger.LogTeardownDeleteSubscription(Name, FullyQualifiedNamespace, TopicName);
+                            await _client.DeleteSubscriptionAsync(TopicName, Name).ConfigureAwait(false);
+                        }));
+                    }
+                    else
+                    {
+                        _disposables.AddRange(_rules.Select(r => AsyncDisposable.Create(async () =>
+                        {
+                            if (await _client.RuleExistsAsync(TopicName, Name, r.Name).ConfigureAwait(false))
+                            {
+                                _logger.LogTeardownDeleteSubscriptionRule(r.Name, FullyQualifiedNamespace, TopicName, Name);
+                                await _client.DeleteRuleAsync(TopicName, Name, r.Name).ConfigureAwait(false);
+                            }
+                        })));
+                    }
+                }
+
+                GC.SuppressFinalize(this);
+            }
         }
+    }
+
+    internal static partial class TempTopicSubscriptionILoggerExtensions
+    {
+        private const LogLevel SetupTeardownLogLevel = LogLevel.Debug;
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Setup] Create new Azure Service Bus topic subscription '{SubscriptionName}' in '{Namespace}/{TopicName}'")]
+        internal static partial void LogSetupCreateNewSubscription(this ILogger logger, string subscriptionName, string @namespace, string topicName);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Setup] Use already existing Azure Service Bus topic subscription '{SubscriptionName}' in '{Namespace}/{TopicName}'")]
+        internal static partial void LogSetupUseExistingSubscription(this ILogger logger, string subscriptionName, string @namespace, string topicName);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test] Create new Azure Service Bus topic subscription rule '{RuleName}' in '{Namespace}/{TopicName}/{SubscriptionName}'")]
+        internal static partial void LogCreateSubscriptionRule(this ILogger logger, string ruleName, string @namespace, string topicName, string subscriptionName);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test] Skip creation of Azure Service Bus topic subscription rule '{RuleName}' in '{Namespace}/{TopicName}/{SubscriptionName}' as it already exists")]
+        internal static partial void LogSkipCreateSubscriptionRule(this ILogger logger, string ruleName, string @namespace, string topicName, string subscriptionName);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Teardown] Delete Azure Service Bus topic subscription rule '{RuleName}' in {Namespace}/{TopicName}/{SubscriptionName}")]
+        internal static partial void LogTeardownDeleteSubscriptionRule(this ILogger logger, string ruleName, string @namespace, string topicName, string subscriptionName);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Teardown] Delete Azure Service Bus topic subscription '{SubscriptionName}' in {Namespace}/{TopicName}")]
+        internal static partial void LogTeardownDeleteSubscription(this ILogger logger, string subscriptionName, string @namespace, string topicName);
     }
 }
