@@ -6,8 +6,9 @@ namespace Arcus.Testing
     /// <summary>
     /// Represents the root exception for any test assertions failure in the library.
     /// </summary>
-    [Serializable]
+#pragma warning disable S3925 // Custom exceptions are serializable by default in modern .NET.
     public class AssertionException : Exception
+#pragma warning restore S3925
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="AssertionException" /> class.
