@@ -80,9 +80,7 @@ namespace Arcus.Testing
         /// <exception cref="FileNotFoundException">
         ///     Thrown when there exists no test resource file in the current test resource directory with the given <paramref name="fileName"/>.
         /// </exception>
-#pragma warning disable S1133 // Should wait until the next major version to remove obsolete code.
         [Obsolete("Will be removed in v3.0, please use " + nameof(ReadFileText) + " instead")]
-#pragma warning restore S1133
         public string ReadFileTextByName(string fileName)
         {
             return ReadFileText(fileName);
@@ -100,9 +98,7 @@ namespace Arcus.Testing
         /// <exception cref="FileNotFoundException">
         ///     Thrown when there exists no test resource file in the current test resource directory with the given <paramref name="searchPattern"/>.
         /// </exception>
-#pragma warning disable S1133 // Should wait until the next major version to remove obsolete code.
         [Obsolete("Will be removed in v3.0, please use " + nameof(ReadFileText) + " instead")]
-#pragma warning restore S1133
         public string ReadFileTextByPattern(string searchPattern)
         {
             return ReadFileText(searchPattern);
@@ -142,9 +138,7 @@ namespace Arcus.Testing
         /// <exception cref="FileNotFoundException">
         ///     Thrown when there exists no test resource file in the current test resource directory with the given <paramref name="fileName"/>.
         /// </exception>
-#pragma warning disable S1133 // Should wait until the next major version to remove obsolete code.
         [Obsolete("Will be removed in v3.0, please use " + nameof(ReadFileBytes) + " instead")]
-#pragma warning restore S1133
         public byte[] ReadFileBytesByName(string fileName)
         {
             return ReadFileBytes(fileName);
@@ -162,9 +156,7 @@ namespace Arcus.Testing
         /// <exception cref="FileNotFoundException">
         ///     Thrown when there exists no test resource file in the current test resource directory with the given <paramref name="searchPattern"/>.
         /// </exception>
-#pragma warning disable S1133 // Should wait until the next major version to remove obsolete code.
         [Obsolete("Will be removed in v3.0, please use " + nameof(ReadFileBytes) + " instead")]
-#pragma warning restore S1133
         public byte[] ReadFileBytesByPattern(string searchPattern)
         {
             return ReadFileBytes(searchPattern);
@@ -238,7 +230,7 @@ namespace Arcus.Testing
                 return false;
             }
 
-            return Path.FullName.Equals(other.Path.FullName, StringComparison.InvariantCultureIgnoreCase);
+            return Path.FullName.Equals(other.Path.FullName, StringComparison.OrdinalIgnoreCase);
         }
 
         /// <summary>
@@ -292,7 +284,9 @@ namespace Arcus.Testing
         /// <exception cref="DirectoryNotFoundException">
         ///     Thrown when no test resource subdirectory is found within the current test resource directory with the given <paramref name="subDirectoryName"/>.
         /// </exception>
+#pragma warning disable CA2225 // Does not require public 'Divide' alternative, as this is the .WithSubDirectory(...) method.
         public static ResourceDirectory operator /(ResourceDirectory current, string subDirectoryName)
+#pragma warning restore CA2225
         {
             ArgumentNullException.ThrowIfNull(current);
             return current.WithSubDirectory(subDirectoryName);
