@@ -63,9 +63,9 @@ namespace Arcus.Testing
         /// <exception cref="ArgumentException">Thrown when the <paramref name="blobContainerUri"/> or the <paramref name="blobName"/> is blank.</exception>
         /// <exception cref="ArgumentNullException">Thrown when <paramref name="blobContainerUri"/> or the <paramref name="blobContent"/> is <c>null</c>.</exception>
         [Obsolete("Will be removed in v3.0, please use the " + nameof(UpsertFileAsync) + " instead which provides the exact same functionality")]
-        public static async Task<TemporaryBlobFile> UploadIfNotExistsAsync(Uri blobContainerUri, string blobName, BinaryData blobContent, ILogger logger)
+        public static Task<TemporaryBlobFile> UploadIfNotExistsAsync(Uri blobContainerUri, string blobName, BinaryData blobContent, ILogger logger)
         {
-            return await UpsertFileAsync(blobContainerUri, blobName, blobContent, logger);
+            return UpsertFileAsync(blobContainerUri, blobName, blobContent, logger);
         }
 
         /// <summary>
@@ -76,9 +76,9 @@ namespace Arcus.Testing
         /// <param name="logger">The logger to write diagnostic messages during the upload process.</param>
         /// <exception cref="ArgumentNullException">Thrown when the <paramref name="blobClient"/> or the <paramref name="blobContent"/> is <c>null</c>.</exception>
         [Obsolete("Will be removed in v3.0, please use the " + nameof(UpsertFileAsync) + " instead which provides the exact same functionality")]
-        public static async Task<TemporaryBlobFile> UploadIfNotExistsAsync(BlobClient blobClient, BinaryData blobContent, ILogger logger)
+        public static Task<TemporaryBlobFile> UploadIfNotExistsAsync(BlobClient blobClient, BinaryData blobContent, ILogger logger)
         {
-            return await UpsertFileAsync(blobClient, blobContent, logger);
+            return UpsertFileAsync(blobClient, blobContent, logger);
         }
 
         /// <summary>
@@ -98,7 +98,7 @@ namespace Arcus.Testing
             ArgumentNullException.ThrowIfNull(blobContent);
             logger ??= NullLogger.Instance;
 
-            (bool createdByUs, BinaryData originalData) = await EnsureBlobContentCreatedAsync(blobClient, blobContent, logger);
+            (bool createdByUs, BinaryData originalData) = await EnsureBlobContentCreatedAsync(blobClient, blobContent, logger).ConfigureAwait(false);
 
             return new TemporaryBlobFile(blobClient, createdByUs, originalData, logger);
         }
@@ -119,7 +119,7 @@ namespace Arcus.Testing
         /// <param name="logger">The logger to write diagnostic messages during the upload process.</param>
         /// <exception cref="ArgumentException">Thrown when the <paramref name="blobContainerUri"/> or the <paramref name="blobName"/> is blank.</exception>
         /// <exception cref="ArgumentNullException">Thrown when <paramref name="blobContainerUri"/> or the <paramref name="blobContent"/> is <c>null</c>.</exception>
-        public static async Task<TemporaryBlobFile> UpsertFileAsync(Uri blobContainerUri, string blobName, BinaryData blobContent, ILogger logger)
+        public static Task<TemporaryBlobFile> UpsertFileAsync(Uri blobContainerUri, string blobName, BinaryData blobContent, ILogger logger)
         {
             ArgumentNullException.ThrowIfNull(blobContainerUri);
             ArgumentException.ThrowIfNullOrWhiteSpace(blobName);
@@ -127,7 +127,7 @@ namespace Arcus.Testing
             var containerClient = new BlobContainerClient(blobContainerUri, new DefaultAzureCredential());
             BlobClient blobClient = containerClient.GetBlobClient(blobName);
 
-            return await UpsertFileAsync(blobClient, blobContent, logger);
+            return UpsertFileAsync(blobClient, blobContent, logger);
         }
 
         private static async Task<(bool createdByUs, BinaryData originalData)> EnsureBlobContentCreatedAsync(
@@ -135,18 +135,18 @@ namespace Arcus.Testing
             BinaryData newContent,
             ILogger logger)
         {
-            if (await client.ExistsAsync())
+            if (await client.ExistsAsync().ConfigureAwait(false))
             {
-                BlobDownloadResult originalContent = await client.DownloadContentAsync();
+                BlobDownloadResult originalContent = await client.DownloadContentAsync().ConfigureAwait(false);
 
-                logger.LogDebug("[Test:Setup] Replace already existing Azure Blob file '{BlobName}' in container '{AccountName}/{ContainerName}'", client.Name, client.AccountName, client.BlobContainerName);
-                await client.UploadAsync(newContent, overwrite: true);
+                logger.LogSetupReplaceFile(client.Name, client.AccountName, client.BlobContainerName);
+                await client.UploadAsync(newContent, overwrite: true).ConfigureAwait(false);
 
                 return (createdByUs: false, originalContent.Content);
             }
 
-            logger.LogDebug("[Test:Setup] Upload Azure Blob file '{BlobName}' to container '{AccountName}/{ContainerName}'", client.Name, client.AccountName, client.BlobContainerName);
-            await client.UploadAsync(newContent);
+            logger.LogSetupUploadNewFile(client.Name, client.AccountName, client.BlobContainerName);
+            await client.UploadAsync(newContent).ConfigureAwait(false);
 
             return (createdByUs: true, originalData: null);
         }
@@ -159,16 +159,41 @@ namespace Arcus.Testing
         {
             if (_createdByUs)
             {
-                _logger.LogDebug("[Test:Teardown] Delete Azure Blob file '{BlobName}' from container '{AccountName}/{ContainerName}'", Client.Name, Client.AccountName, Client.BlobContainerName);
-                await Client.DeleteIfExistsAsync();
+                _logger.LogTeardownDeleteFile(Client.Name, Client.AccountName, Client.BlobContainerName);
+                await Client.DeleteIfExistsAsync().ConfigureAwait(false);
             }
             else if (_originalData != null)
             {
-                _logger.LogDebug("[Test:Teardown] Revert replaced Azure Blob file '{BlobName}' to original content in container '{AccountName}/{ContainerName}'", Client.Name, Client.AccountName, Client.BlobContainerName);
-                await Client.UploadAsync(_originalData, overwrite: true);
+                _logger.LogTeardownRevertFile(Client.Name, Client.AccountName, Client.BlobContainerName);
+                await Client.UploadAsync(_originalData, overwrite: true).ConfigureAwait(false);
             }
 
             GC.SuppressFinalize(this);
         }
+    }
+
+    internal static partial class TempBlobFileILoggerExtensions
+    {
+        private const LogLevel SetupTeardownLogLevel = LogLevel.Debug;
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Setup] Upload Azure Blob file '{BlobName}' to container '{AccountName}/{ContainerName}'")]
+        internal static partial void LogSetupUploadNewFile(this ILogger logger, string blobName, string accountName, string containerName);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Setup] Replace already existing Azure Blob file '{BlobName}' in container '{AccountName}/{ContainerName}'")]
+        internal static partial void LogSetupReplaceFile(this ILogger logger, string blobName, string accountName, string containerName);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Teardown] Delete Azure Blob file '{BlobName}' from container '{AccountName}/{ContainerName}'")]
+        internal static partial void LogTeardownDeleteFile(this ILogger logger, string blobName, string accountName, string containerName);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Teardown] Revert replaced Azure Blob file '{BlobName}' to original content in container '{AccountName}/{ContainerName}'")]
+        internal static partial void LogTeardownRevertFile(this ILogger logger, string blobName, string accountName, string containerName);
     }
 }
