@@ -13,14 +13,16 @@ using static Arcus.Testing.NoSqlExtraction;
 namespace Arcus.Testing
 {
     /// <summary>
-    /// Represents a temporary Azure Cosmos NoSql document that will be deleted after the instance is disposed.
+    /// Represents a temporary Azure Cosmos DB for NoSQL document that will be deleted after the instance is disposed.
     /// </summary>
     public class TemporaryNoSqlItem : IAsyncDisposable
     {
         private readonly Container _container;
         private readonly Type _itemType;
         private readonly bool _createdByUs;
+#pragma warning disable CA2213 // Disposable field is disposed via the disposable collection.
         private readonly Stream _originalItemStream;
+#pragma warning restore CA2213
         private readonly ILogger _logger;
 
         private TemporaryNoSqlItem(
@@ -47,7 +49,7 @@ namespace Arcus.Testing
         }
 
         /// <summary>
-        /// Gets the unique identifier of this temporary NoSql item.
+        /// Gets the unique identifier of this temporary NoSQL item.
         /// </summary>
         public string Id { get; }
 
@@ -57,92 +59,90 @@ namespace Arcus.Testing
         public PartitionKey PartitionKey { get; }
 
         /// <summary>
-        /// Creates a temporary item to an Azure Cosmos NoSql container.
+        /// Creates a temporary item to an Azure Cosmos DB for NoSQL container.
         /// </summary>
-        /// <param name="container">The NoSql container where a temporary item should be created.</param>
-        /// <param name="item">The item to temporary create in the NoSql container.</param>
+        /// <param name="container">The NoSQL container where a temporary item should be created.</param>
+        /// <param name="item">The item to temporary create in the NoSQL container.</param>
         /// <param name="logger">The logger instance to write diagnostic information during the lifetime of the test fixture.</param>
         /// <exception cref="ArgumentNullException">Thrown when the <paramref name="container"/> is <c>null</c>.</exception>
         [Obsolete("Will be removed in v3.0, please use the " + nameof(UpsertItemAsync) + " instead which provides the exact same functionality")]
-        public static async Task<TemporaryNoSqlItem> InsertIfNotExistsAsync<TItem>(Container container, TItem item, ILogger logger)
+        public static Task<TemporaryNoSqlItem> InsertIfNotExistsAsync<TItem>(Container container, TItem item, ILogger logger)
         {
-            return await UpsertItemAsync(container, item, logger);
+            return UpsertItemAsync(container, item, logger);
         }
 
         /// <summary>
-        /// Creates a new or replaces an existing NoSql item in an Azure Cosmos NoSql container.
+        /// Creates a new or replaces an existing NoSQL item in an Azure Cosmos DB for NoSQL container.
         /// </summary>
         /// <remarks>
         ///     ⚡ Item will be deleted (if new) or reverted (if existing) when the <see cref="TemporaryNoSqlItem"/> is disposed.
         /// </remarks>
-        /// <param name="container">The NoSql container where a temporary item should be created.</param>
-        /// <param name="item">The item to temporary create in the NoSql container.</param>
+        /// <param name="container">The NoSQL container where a temporary item should be created.</param>
+        /// <param name="item">The item to temporary create in the NoSQL container.</param>
         /// <param name="logger">The logger instance to write diagnostic information during the lifetime of the test fixture.</param>
         /// <exception cref="ArgumentNullException">Thrown when the <paramref name="container"/> is <c>null</c>.</exception>
         public static async Task<TemporaryNoSqlItem> UpsertItemAsync<TItem>(Container container, TItem item, ILogger logger)
         {
             ArgumentNullException.ThrowIfNull(container);
             logger ??= NullLogger.Instance;
+            Type itemType = typeof(TItem);
 
             JsonNode json = JsonSerializer.SerializeToNode(item, SerializeToNodeOptions);
-            string itemId = ExtractIdFromItem(json, typeof(TItem));
+            string itemId = ExtractIdFromItem(json, itemType);
 
-            ContainerResponse resp = await container.ReadContainerAsync();
+            ContainerResponse resp = await container.ReadContainerAsync().ConfigureAwait(false);
             PartitionKey partitionKey = ExtractPartitionKeyFromItem(json, resp.Resource);
 
             try
             {
-                return await ReplaceExistingItemAsync(container, itemId, partitionKey, item, logger);
+                Stream originalItemStream = await ReplaceExistingItemOnSetupAsync(container, itemType, itemId, partitionKey, item, logger).ConfigureAwait(false);
+                return new TemporaryNoSqlItem(container, itemId, partitionKey, itemType, createdByUs: false, originalItemStream, logger);
             }
             catch (CosmosException exception) when (exception.StatusCode is HttpStatusCode.NotFound)
             {
-                return await InsertNewItemAsync(container, itemId, partitionKey, item, logger);
+                await InsertNewItemOnSetupAsync(container, itemType, itemId, item, logger).ConfigureAwait(false);
+                return new TemporaryNoSqlItem(container, itemId, partitionKey, itemType, createdByUs: true, originalItemStream: null, logger);
             }
         }
 
-        private static async Task<TemporaryNoSqlItem> ReplaceExistingItemAsync<TItem>(
+        private static async Task<Stream> ReplaceExistingItemOnSetupAsync<TItem>(
             Container container,
+            Type itemType,
             string itemId,
             PartitionKey partitionKey,
             TItem item,
             ILogger logger)
         {
-            logger.LogTrace("[Test:Setup] Try replacing Azure Cosmos NoSql '{ItemType}' item '{ItemId}' in container '{DatabaseName}/{ContainerName}'...", typeof(TItem).Name, itemId, container.Database.Id, container.Id);
-            ItemResponse<TItem> response = await container.ReadItemAsync<TItem>(itemId, partitionKey);
+            logger.LogBeforeSetupReplaceItem(itemType.Name, itemId, container.Database.Id, container.Id);
+            ItemResponse<TItem> response = await container.ReadItemAsync<TItem>(itemId, partitionKey).ConfigureAwait(false);
 
             CosmosSerializer serializer = container.Database.Client.ClientOptions.Serializer;
             if (serializer is null)
             {
                 throw new InvalidOperationException(
-                    "[Test:Setup] Cannot temporary insert an Azure Cosmos NoSql item in a NoSql container because no JSON serializer was set in the Cosmos client options");
+                    $"[Test:Setup] Cannot temporary insert an NoSQL item in the Azure Cosmos DB for NoSQL container '{container.Database.Id}/{container.Id}' because no JSON serializer was set in the Cosmos DB client options");
             }
 
             TItem original = response.Resource;
             var originalItemStream = serializer.ToStream(original);
 
-            logger.LogDebug("[Test:Setup] Replace Azure Cosmos NoSql '{ItemType}' item '{ItemId}' in container '{DatabaseName}/{ContainerName}'", typeof(TItem).Name, itemId, container.Database.Id, container.Id);
-            await container.ReplaceItemAsync(item, itemId, partitionKey);
+            logger.LogSetupReplaceItem(itemType.Name, itemId, container.Database.Id, container.Id);
+            await container.ReplaceItemAsync(item, itemId, partitionKey).ConfigureAwait(false);
 
-            return new TemporaryNoSqlItem(container, itemId, partitionKey, typeof(TItem), createdByUs: false, originalItemStream, logger);
+            return originalItemStream;
         }
 
-        private static async Task<TemporaryNoSqlItem> InsertNewItemAsync<TItem>(
-            Container container,
-            string itemId,
-            PartitionKey partitionKey,
-            TItem item,
-            ILogger logger)
+        private static async Task InsertNewItemOnSetupAsync<TItem>(Container container, Type itemType, string itemId, TItem item, ILogger logger)
         {
-            logger.LogDebug("[Test:Setup] Insert new Azure Cosmos NoSql '{ItemType}' item '{ItemId}' to container '{DatabaseName}/{ContainerName}'", typeof(TItem).Name, itemId, container.Database.Id, container.Id);
-            ItemResponse<TItem> response = await container.CreateItemAsync(item);
+            logger.LogSetupInsertNewItem(itemType.Name, itemId, container.Database.Id, container.Id);
+            ItemResponse<TItem> response = await container.CreateItemAsync(item).ConfigureAwait(false);
 
             if (response.StatusCode != HttpStatusCode.Created)
             {
-                throw new RequestFailedException(
-                    (int) response.StatusCode, $"[Test:Setup] Unable to insert a new Azure Cosmos NoSql '{typeof(TItem).Name}' item into the container '{container.Database.Id}/{container.Id}': {response.StatusCode} - " + response.Diagnostics);
+                throw new RequestFailedException((int) response.StatusCode,
+                    $"[Test:Setup] Failed to insert a new NoSQL '{itemType.Name}' item into the Azure Cosmos DB for NoSQL container '{container.Database.Id}/{container.Id}' " +
+                    $"since the create operation responded with a failure: {(int) response.StatusCode} {response.StatusCode}: {response.Diagnostics}");
             }
-
-            return new TemporaryNoSqlItem(container, itemId, partitionKey, typeof(TItem), createdByUs: true, originalItemStream: null, logger);
         }
 
         /// <summary>
@@ -151,44 +151,79 @@ namespace Arcus.Testing
         /// <returns>A task that represents the asynchronous dispose operation.</returns>
         public async ValueTask DisposeAsync()
         {
-            await using var disposables = new DisposableCollection(_logger);
-
-            if (_createdByUs && _originalItemStream is null)
+            var disposables = new DisposableCollection(_logger);
+            await using (disposables.ConfigureAwait(false))
             {
-                disposables.Add(AsyncDisposable.Create(async () =>
+                if (_createdByUs && _originalItemStream is null)
                 {
-                    _logger.LogDebug("[Test:Teardown] Delete Azure Cosmos NoSql '{ItemType}' item '{ItemId}' in container '{DatabaseName}/{ContainerName}'", _itemType.Name, Id, _container.Database.Id, _container.Id);
-                    using ResponseMessage response = await _container.DeleteItemStreamAsync(Id, PartitionKey);
+                    disposables.Add(AsyncDisposable.Create(DeleteItemOnTeardownAsync));
+                }
+                else
+                {
+                    disposables.Add(AsyncDisposable.Create(RevertItemOnTeardownAsync));
+                    disposables.Add((IAsyncDisposable) _originalItemStream);
+                }
 
-                    if (!response.IsSuccessStatusCode && response.StatusCode != HttpStatusCode.NotFound)
-                    {
-                        throw new RequestFailedException(
-                            $"[Test:Teardown] Failed to delete Azure Cosmos NoSql '{_itemType.Name}' item '{Id}' {PartitionKey} in container '{_container.Database.Id}/{_container.Id}' " +
-                            $"since the delete operation responded with a failure: {(int) response.StatusCode} {response.StatusCode}: {response.ErrorMessage}");
-                    }
-                }));
+                GC.SuppressFinalize(this);
             }
-            else
-            {
-                disposables.Add(AsyncDisposable.Create(async () =>
-                {
-                    _logger.LogDebug("[Test:Teardown] Revert replaced Azure Cosmos NoSql '{ItemType}' item '{ItemId}' in container '{DatabaseName}/{ContainerName}'", _itemType.Name, Id, _container.Database.Id, _container.Id);
-                    using ResponseMessage response = await _container.ReplaceItemStreamAsync(_originalItemStream, Id, PartitionKey);
-
-                    if (!response.IsSuccessStatusCode)
-                    {
-                        throw new RequestFailedException(
-                            $"[Test:Teardown] Failed to revert Azure Cosmos NoSql '{_itemType.Name}' item '{Id}' {PartitionKey} in container '{_container.Database.Id}/{_container.Id}' " +
-                            $"since the replace operation responded with a failure: {(int) response.StatusCode} {response.StatusCode}: {response.ErrorMessage}");
-                    }
-                }));
-                disposables.Add(AsyncDisposable.Create(async () =>
-                {
-                    await _originalItemStream.DisposeAsync();
-                }));
-            }
-
-            GC.SuppressFinalize(this);
         }
+
+        private async Task DeleteItemOnTeardownAsync()
+        {
+            _logger.LogTeardownDeleteItem(_itemType.Name, Id, _container.Database.Id, _container.Id);
+            using ResponseMessage response = await _container.DeleteItemStreamAsync(Id, PartitionKey).ConfigureAwait(false);
+
+            if (!response.IsSuccessStatusCode && response.StatusCode != HttpStatusCode.NotFound)
+            {
+                throw new RequestFailedException(
+                    (int) response.StatusCode,
+                    $"[Test:Teardown] Failed to delete NoSQL '{_itemType.Name}' item '{Id}' {PartitionKey} in Azure Cosmos DB for NoSQL container '{_container.Database.Id}/{_container.Id}' " +
+                    $"since the delete operation responded with a failure: {(int) response.StatusCode} {response.StatusCode}: {response.ErrorMessage}");
+            }
+        }
+
+        private async Task RevertItemOnTeardownAsync()
+        {
+            _logger.LogTeardownRevertItem(_itemType.Name, Id, _container.Database.Id, _container.Id);
+            using ResponseMessage response = await _container.ReplaceItemStreamAsync(_originalItemStream, Id, PartitionKey).ConfigureAwait(false);
+
+            if (!response.IsSuccessStatusCode)
+            {
+                throw new RequestFailedException(
+                    (int) response.StatusCode,
+                    $"[Test:Teardown] Failed to revert NoSQL '{_itemType.Name}' item '{Id}' {PartitionKey} in Azure Cosmos DB for NoSQL container '{_container.Database.Id}/{_container.Id}' " +
+                    $"since the replace operation responded with a failure: {(int) response.StatusCode} {response.StatusCode}: {response.ErrorMessage}");
+            }
+        }
+    }
+
+    internal static partial class TempNoSqlItemILoggerExtensions
+    {
+        private const LogLevel SetupTeardownLogLevel = LogLevel.Debug;
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Setup] Insert new NoSQL '{ItemType}' item '{ItemId}' to Azure Cosmos DB for NoSQL container '{DatabaseName}/{ContainerName}'")]
+        internal static partial void LogSetupInsertNewItem(this ILogger logger, string itemType, string itemId, string databaseName, string containerName);
+
+        [LoggerMessage(
+            Level = LogLevel.Trace,
+            Message = "[Test:Setup] Try replacing NoSQL '{ItemType}' item '{ItemId}' in Azure Cosmos DB For NoSQL container '{DatabaseName}/{ContainerName}'...")]
+        internal static partial void LogBeforeSetupReplaceItem(this ILogger logger, string itemTYpe, string itemId, string databaseName, string containerName);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Setup] Replace NoSQL '{ItemType}' item '{ItemId}' in Azure Cosmos DB for NoSQL container '{DatabaseName}/{ContainerName}'")]
+        internal static partial void LogSetupReplaceItem(this ILogger logger, string itemType, string itemId, string databaseName, string containerName);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Teardown] Delete NoSQL '{ItemType}' item '{ItemId}' in Azure Cosmos DB for NoSQL container '{DatabaseName}/{ContainerName}'")]
+        internal static partial void LogTeardownDeleteItem(this ILogger logger, string itemType, string itemId, string databaseName, string containerName);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Teardown] Revert replaced NoSQL '{ItemType}' item '{ItemId}' in Azure Cosmos DB for NoSQL container '{DatabaseName}/{ContainerName}'")]
+        internal static partial void LogTeardownRevertItem(this ILogger logger, string itemType, string itemId, string databaseName, string containerName);
     }
 }
