@@ -151,8 +151,10 @@ namespace Arcus.Testing
     public class TemporaryShareDirectory : IAsyncDisposable
     {
         private readonly bool _createdByUs;
+        private readonly ShareDirectoryClient _client;
         private readonly Collection<TemporaryShareFile> _files = [];
         private readonly TemporaryShareDirectoryOptions _options;
+        private readonly DisposableCollection _disposables;
         private readonly ILogger _logger;
 
         private TemporaryShareDirectory(
@@ -166,19 +168,36 @@ namespace Arcus.Testing
             _createdByUs = createdByUs;
             _options = options;
             _logger = logger ?? NullLogger.Instance;
+            _disposables = new DisposableCollection(_logger);
 
-            Client = client;
+            _client = client;
         }
 
         /// <summary>
         /// Represents the client to interact with the temporary stored Azure Files share currently in storage.
         /// </summary>
-        public ShareDirectoryClient Client { get; }
+        /// <exception cref="ObjectDisposedException">Thrown when the test fixture was already teared down.</exception>
+        public ShareDirectoryClient Client
+        {
+            get
+            {
+                ObjectDisposedException.ThrowIf(_disposables.IsDisposed, this);
+                return _client;
+            }
+        }
 
         /// <summary>
         /// Gets the options to manipulate the deletion of the <see cref="TemporaryShareDirectory"/>.
         /// </summary>
-        public OnTeardownTemporaryShareDirectoryOptions OnTeardown => _options.OnTeardown;
+        /// <exception cref="ObjectDisposedException">Thrown when the test fixture was already teared down.</exception>
+        public OnTeardownTemporaryShareDirectoryOptions OnTeardown
+        {
+            get
+            {
+                ObjectDisposedException.ThrowIf(_disposables.IsDisposed, this);
+                return _options.OnTeardown;
+            }
+        }
 
         /// <summary>
         /// Creates a temporary directory share on an Azure Files share resource.
@@ -188,9 +207,9 @@ namespace Arcus.Testing
         /// <param name="logger">The logger instance to write diagnostic traces during the lifetime of the fixture.</param>
         /// <exception cref="ArgumentNullException">Thrown when the <paramref name="shareClient"/> is <c>null</c>.</exception>
         /// <exception cref="ArgumentException">Thrown when the <paramref name="directoryName"/> is blank.</exception>
-        public static async Task<TemporaryShareDirectory> CreateIfNotExistsAsync(ShareClient shareClient, string directoryName, ILogger logger)
+        public static Task<TemporaryShareDirectory> CreateIfNotExistsAsync(ShareClient shareClient, string directoryName, ILogger logger)
         {
-            return await CreateIfNotExistsAsync(shareClient, directoryName, logger, configureOptions: null);
+            return CreateIfNotExistsAsync(shareClient, directoryName, logger, configureOptions: null);
         }
 
         /// <summary>
@@ -202,7 +221,7 @@ namespace Arcus.Testing
         /// <param name="configureOptions">The additional options to manipulate the behavior of the test fixture during its lifetime.</param>
         /// <exception cref="ArgumentNullException">Thrown when the <paramref name="shareClient"/> is <c>null</c>.</exception>
         /// <exception cref="ArgumentException">Thrown when the <paramref name="directoryName"/> is blank.</exception>
-        public static async Task<TemporaryShareDirectory> CreateIfNotExistsAsync(
+        public static Task<TemporaryShareDirectory> CreateIfNotExistsAsync(
             ShareClient shareClient,
             string directoryName,
             ILogger logger,
@@ -212,7 +231,7 @@ namespace Arcus.Testing
             ArgumentException.ThrowIfNullOrWhiteSpace(directoryName);
 
             ShareDirectoryClient dirClient = shareClient.GetDirectoryClient(directoryName);
-            return await CreateIfNotExistsAsync(dirClient, logger, configureOptions);
+            return CreateIfNotExistsAsync(dirClient, logger, configureOptions);
         }
 
         /// <summary>
@@ -221,9 +240,9 @@ namespace Arcus.Testing
         /// <param name="directoryClient">The client to interact with the directory share in the Azure Files share resource.</param>
         /// <param name="logger">The logger instance to write diagnostic traces during the lifetime of the fixture.</param>
         /// <exception cref="ArgumentNullException">Thrown when the <paramref name="directoryClient"/> is <c>null</c>.</exception>
-        public static async Task<TemporaryShareDirectory> CreateIfNotExistsAsync(ShareDirectoryClient directoryClient, ILogger logger)
+        public static Task<TemporaryShareDirectory> CreateIfNotExistsAsync(ShareDirectoryClient directoryClient, ILogger logger)
         {
-            return await CreateIfNotExistsAsync(directoryClient, logger, configureOptions: null);
+            return CreateIfNotExistsAsync(directoryClient, logger, configureOptions: null);
         }
 
         /// <summary>
@@ -244,18 +263,18 @@ namespace Arcus.Testing
             var options = new TemporaryShareDirectoryOptions();
             configureOptions?.Invoke(options);
 
-            if (await directoryClient.ExistsAsync())
+            if (await directoryClient.ExistsAsync().ConfigureAwait(false))
             {
-                logger.LogTrace("[Test:Setup] Use already existing Azure Files share directory '{DirectoryName}' at '{AccountName}/{DirectoryPath}'", directoryClient.Name, directoryClient.AccountName, directoryClient.Path);
-                await CleanDirectoryOnSetupAsync(directoryClient, options, logger);
+                logger.LogSetupUseExistingDirectory(directoryClient.Name, directoryClient.AccountName, directoryClient.Path);
+                await CleanDirectoryOnSetupAsync(directoryClient, options, logger).ConfigureAwait(false);
 
                 return new TemporaryShareDirectory(directoryClient, createdByUs: false, options, logger);
             }
 
             try
             {
-                logger.LogTrace("[Test:Setup] Create new Azure Files share directory '{DirectoryName}' at '{AccountName}/{DirectoryPath}'", directoryClient.Name, directoryClient.AccountName, directoryClient.Path);
-                await directoryClient.CreateAsync();
+                logger.LogSetupCreateNewDirectory(directoryClient.Name, directoryClient.AccountName, directoryClient.Path);
+                await directoryClient.CreateAsync().ConfigureAwait(false);
             }
             catch (RequestFailedException exception) when (exception.ErrorCode == ShareErrorCode.ShareNotFound)
             {
@@ -278,11 +297,11 @@ namespace Arcus.Testing
 
             if (options.OnSetup.Items is OnSetupDirectoryShare.CleanIfExisted)
             {
-                await DeleteAllDirectoryContentsAsync(directoryClient, TestFixture.Setup, logger);
+                await DeleteAllDirectoryContentsAsync(directoryClient, TestFixture.Setup, logger).ConfigureAwait(false);
             }
             else if (options.OnSetup.Items is OnSetupDirectoryShare.CleanIfMatched)
             {
-                await DeleteDirectoryContentsAsync(directoryClient, options.OnSetup.IsMatch, TestFixture.Setup, logger);
+                await DeleteDirectoryContentsAsync(directoryClient, options.OnSetup.IsMatch, TestFixture.Setup, logger).ConfigureAwait(false);
             }
         }
 
@@ -294,11 +313,13 @@ namespace Arcus.Testing
         /// </remarks>
         /// <param name="fileName">The name of the file to upload to the share directory.</param>
         /// <param name="fileContents">The contents of the file to upload to the share directory.</param>
+        /// <exception cref="ObjectDisposedException">Thrown when the test fixture was already teared down.</exception>
         /// <exception cref="ArgumentException">Thrown when the <paramref name="fileName"/> is blank.</exception>
         /// <exception cref="ArgumentNullException">Thrown when the <paramref name="fileContents"/> is <c>null</c>.</exception>
         public async Task UpsertFileAsync(string fileName, Stream fileContents)
         {
-            _files.Add(await TemporaryShareFile.UpsertFileAsync(Client, fileName, fileContents, _logger));
+            ObjectDisposedException.ThrowIf(_disposables.IsDisposed, this);
+            _files.Add(await TemporaryShareFile.UpsertFileAsync(Client, fileName, fileContents, _logger).ConfigureAwait(false));
         }
 
         /// <summary>
@@ -307,22 +328,29 @@ namespace Arcus.Testing
         /// <returns>A task that represents the asynchronous dispose operation.</returns>
         public async ValueTask DisposeAsync()
         {
-            await using var disposables = new DisposableCollection(_logger);
-            disposables.AddRange(_files);
-
-            if (_createdByUs)
+            if (_disposables.IsDisposed)
             {
-                await DeleteAllDirectoryContentsAsync(Client, TestFixture.Teardown, _logger);
-
-                disposables.Add(AsyncDisposable.Create(async () =>
-                {
-                    _logger.LogTrace("[Test:Teardown] Delete Azure Files share directory '{DirectoryName}' at '{AccountName}/{DirectoryPath}'", Client.Name, Client.AccountName, Client.Path);
-                    await Client.DeleteAsync();
-                }));
+                return;
             }
-            else
+
+            await using (_disposables.ConfigureAwait(false))
             {
-                await CleanDirectoryOnTeardownAsync();
+                _disposables.AddRange(_files);
+
+                if (_createdByUs)
+                {
+                    await DeleteAllDirectoryContentsAsync(_client, TestFixture.Teardown, _logger).ConfigureAwait(false);
+
+                    _disposables.Add(AsyncDisposable.Create(async () =>
+                    {
+                        _logger.LogTeardownDeleteDirectory(_client.Name, _client.AccountName, _client.Path);
+                        await _client.DeleteAsync().ConfigureAwait(false);
+                    }));
+                }
+                else
+                {
+                    _disposables.Add(AsyncDisposable.Create(CleanDirectoryOnTeardownAsync));
+                }
             }
 
             GC.SuppressFinalize(this);
@@ -337,11 +365,11 @@ namespace Arcus.Testing
 
             if (_options.OnTeardown.Items is OnTeardownDirectoryShare.CleanAll)
             {
-                await DeleteAllDirectoryContentsAsync(Client, TestFixture.Teardown, _logger);
+                await DeleteAllDirectoryContentsAsync(_client, TestFixture.Teardown, _logger).ConfigureAwait(false);
             }
             else if (_options.OnTeardown.Items is OnTeardownDirectoryShare.CleanIfMatched)
             {
-                await DeleteDirectoryContentsAsync(Client, _options.OnTeardown.IsMatch, TestFixture.Teardown, _logger);
+                await DeleteDirectoryContentsAsync(_client, _options.OnTeardown.IsMatch, TestFixture.Teardown, _logger).ConfigureAwait(false);
             }
         }
 
@@ -353,43 +381,96 @@ namespace Arcus.Testing
             TestFixture state,
             ILogger logger)
         {
-            await using var disposables = new DisposableCollection(logger);
+            await current.ForceCloseAllHandlesAsync(recursive: true).ConfigureAwait(false);
 
-            disposables.Add(AsyncDisposable.Create(async () =>
+            await foreach (ShareFileItem item in current.GetFilesAndDirectoriesAsync().ConfigureAwait(false))
             {
-                await current.ForceCloseAllHandlesAsync(recursive: true);
-
-                await foreach (ShareFileItem item in current.GetFilesAndDirectoriesAsync())
+                if (item.IsDirectory)
                 {
-                    if (item.IsDirectory)
+                    ShareDirectoryClient sub = current.GetSubdirectoryClient(item.Name);
+                    if (shouldDeleteItem(item))
                     {
-                        ShareDirectoryClient sub = current.GetSubdirectoryClient(item.Name);
-                        if (shouldDeleteItem(item))
-                        {
-                            await DeleteAllDirectoryContentsAsync(sub, state, logger);
+                        await DeleteAllDirectoryContentsAsync(sub, state, logger).ConfigureAwait(false);
 
-                            logger.LogTrace("[Test:{State}] Delete Azure Files share directory '{DirectoryName}' at '{AccountName}/{DirectoryPath}'", state, sub.Name, sub.AccountName, sub.Path);
-                            await sub.DeleteIfExistsAsync();
-                        }
-                        else
-                        {
-                            await DeleteDirectoryContentsAsync(sub, shouldDeleteItem, state, logger);
-                        }
+                        LogDeleteDirectory(logger, state, sub.Name, sub.AccountName, sub.Path);
+                        await sub.DeleteIfExistsAsync().ConfigureAwait(false);
                     }
-                    else if (shouldDeleteItem(item))
+                    else
                     {
-                        ShareFileClient file = current.GetFileClient(item.Name);
-
-                        logger.LogTrace("[Test:{State}] Delete Azure Files share item '{FileName}' at '{AccountName}/{FilePath}'", state, file.Name, file.AccountName, file.Path);
-                        await file.DeleteIfExistsAsync();
+                        await DeleteDirectoryContentsAsync(sub, shouldDeleteItem, state, logger).ConfigureAwait(false);
                     }
                 }
-            }));
+                else if (shouldDeleteItem(item))
+                {
+                    ShareFileClient file = current.GetFileClient(item.Name);
+
+                    LogDeleteFile(logger, state, file.Name, file.AccountName, file.Path);
+                    await file.DeleteIfExistsAsync().ConfigureAwait(false);
+                }
+            }
+        }
+
+        private static void LogDeleteDirectory(ILogger logger, TestFixture state, string directoryName, string accountName, string directoryPath)
+        {
+            switch (state)
+            {
+                case TestFixture.Setup:
+                    logger.LogSetupDeleteDirectory(directoryName, accountName, directoryPath);
+                    break;
+
+                case TestFixture.Teardown:
+                    logger.LogTeardownDeleteDirectory(directoryName, accountName, directoryPath);
+                    break;
+            }
+        }
+
+        private static void LogDeleteFile(ILogger logger, TestFixture state, string fileName, string accountName, string filePath)
+        {
+            switch (state)
+            {
+                case TestFixture.Setup:
+                    logger.LogSetupDeleteFile(fileName, accountName, filePath);
+                    break;
+
+                case TestFixture.Teardown:
+                    logger.LogTeardownDeleteFile(fileName, accountName, filePath);
+                    break;
+            }
         }
 
         private static async Task DeleteAllDirectoryContentsAsync(ShareDirectoryClient current, TestFixture state, ILogger logger)
         {
-            await DeleteDirectoryContentsAsync(current, shouldDeleteItem: _ => true, state, logger);
+            await DeleteDirectoryContentsAsync(current, shouldDeleteItem: _ => true, state, logger).ConfigureAwait(false);
         }
+    }
+
+    internal static partial class TempShareDirectoryILoggerExtensions
+    {
+        private const LogLevel SetupTeardownLogLevel = LogLevel.Debug;
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Setup] Create new Azure Files share directory '{DirectoryName}' at '{AccountName}/{DirectoryPath}'")]
+        internal static partial void LogSetupCreateNewDirectory(this ILogger logger, string directoryName, string accountName, string directoryPath);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Setup] Use already existing Azure Files share directory '{DirectoryName}' at '{AccountName}/{DirectoryPath}'")]
+        internal static partial void LogSetupUseExistingDirectory(this ILogger logger, string directoryName, string accountName, string directoryPath);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Setup] Delete Azure Files share directory '{DirectoryName}' at '{AccountName}/{DirectoryPath}'")]
+        internal static partial void LogSetupDeleteDirectory(this ILogger logger, string directoryName, string accountName, string directoryPath);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Setup] Delete Azure Files share item '{FileName}' at '{AccountName}/{FilePath}'")]
+        internal static partial void LogSetupDeleteFile(this ILogger logger, string fileName, string accountName, string filePath);
+
+        [LoggerMessage(
+            Level = SetupTeardownLogLevel,
+            Message = "[Test:Teardown] Delete Azure Files share directory '{DirectoryName}' at '{AccountName}/{DirectoryPath}'")]
+        internal static partial void LogTeardownDeleteDirectory(this ILogger logger, string directoryName, string accountName, string directoryPath);
     }
 }
