@@ -1,6 +1,5 @@
 ﻿using System.Threading.Tasks;
 using Xunit;
-using Xunit.Abstractions;
 
 namespace Arcus.Testing.Tests.Integration.Integration.DataFactory
 {
@@ -22,13 +21,13 @@ namespace Arcus.Testing.Tests.Integration.Integration.DataFactory
         [Fact]
         public async Task StartDebugSession_WithActiveSession_SucceedsByReusingSession()
         {
-            await using (var otherSession = await TemporaryDataFlowDebugSession.StartDebugSessionAsync(DataFactory.ResourceId, Logger, 
+            await using (var otherSession = await TemporaryDataFlowDebugSession.StartDebugSessionAsync(DataFactory.ResourceId, Logger,
                 options => options.ActiveSessionId = _fixture.Value.SessionId))
             {
                 // Assert
                 Assert.Equal(_fixture.Value.SessionId, otherSession.SessionId);
             }
-            
+
             await _fixture.ShouldFindActiveSessionAsync(_fixture.Value.SessionId);
         }
     }
