@@ -5,11 +5,9 @@ using Bogus;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
 using Xunit;
-using Xunit.Abstractions;
 
 namespace Arcus.Testing.Tests.Integration.Storage
 {
-    [Collection(TestCollections.MongoDb)]
     public class TemporaryMongoDbCollectionTests : IntegrationTest
     {
         /// <summary>
@@ -51,7 +49,9 @@ namespace Arcus.Testing.Tests.Integration.Storage
 
             TemporaryMongoDbCollection collection = await WhenTempCollectionCreatedAsync(collectionName);
             Shipment createdByUs = CreateShipment();
+#pragma warning disable CS0618 // Type or member is obsolete: currently still testing deprecated functionality.
             await collection.AddDocumentAsync(createdByUs);
+#pragma warning restore CS0618 // Type or member is obsolete
 
             await context.ShouldStoreCollectionAsync(collectionName);
             await context.ShouldStoreDocumentAsync<Shipment>(collectionName, existingId);
@@ -187,7 +187,9 @@ namespace Arcus.Testing.Tests.Integration.Storage
                 options.OnSetup.CleanMatchingDocuments((Shipment s) => s.BoatName == matchedOnSetup.BoatName);
             });
             collection.OnTeardown.CleanAllDocuments();
+#pragma warning disable CS0618 // Type or member is obsolete: currently still testing deprecated functionality.
             await collection.AddDocumentAsync(createdByUs);
+#pragma warning restore CS0618 // Type or member is obsolete
 
             await context.ShouldNotStoreDocumentAsync<Shipment>(collectionName, matchedOnSetupId);
             await context.ShouldStoreDocumentAsync<Shipment>(collectionName, unmatchedOnSetupId);
