@@ -10,7 +10,6 @@ using Bogus;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using Xunit;
-using Xunit.Abstractions;
 using DirClient = Azure.Storage.Files.Shares.ShareDirectoryClient;
 using FileClient = Azure.Storage.Files.Shares.ShareFileClient;
 
@@ -169,10 +168,17 @@ namespace Arcus.Testing.Tests.Integration.Storage
             await temp.WhenFileFileUploadAsync(file);
             await share.WhenFileAvailableAsync(await share.WhenDirectoryAvailableAsync(temp.Client));
 
-            await temp.DisposeAsync();
+            await WhenTestFixtureTeardownAsync(temp);
 
             await share.ShouldNotHaveFilesAsync(file);
             await share.ShouldNotHaveDirectoriesAsync(dir);
+        }
+
+        private static async Task WhenTestFixtureTeardownAsync(TemporaryShareDirectory temp)
+        {
+            // Calling dispose multiple times should be redundant.
+            await temp.DisposeAsync();
+            await temp.DisposeAsync();
         }
 
         [Fact]
