@@ -15,10 +15,10 @@ namespace Arcus.Testing.Tests.Unit
         /// <returns>An enumerator that can be used to iterate through the collection.</returns>
         public IEnumerator<object[]> GetEnumerator()
         {
-            yield return new object[] {null};
-            yield return new object[] {""};
-            yield return new object[] {" "};
-            yield return new object[] {"        "};
+            yield return new object[] { null };
+            yield return new object[] { "" };
+            yield return new object[] { " " };
+            yield return new object[] { "        " };
             yield return new object[] { Environment.NewLine };
         }
 
